requirements.txt
audioread==3.0.1
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
chatterbox-tts==0.1.3
conformer==0.3.2
Cython==3.1.4
decorator==5.2.1
Deprecated==1.2.18
diffusers==0.29.0
distlib==0.4.0
einops==0.8.1
filelock==3.19.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
identify==2.6.14
idna==3.10
importlib_metadata==8.7.0
jaconv==0.4.0
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
MarkupSafe==3.0.3
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
networkx==3.2.1
nodeenv==1.9.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.19.0
packaging==25.0
pillow==11.3.0
pkuseg==0.0.25
platformdirs==4.4.0
playsound==1.3.0
pooch==1.8.2
pre_commit==4.3.0
protobuf==6.32.1
pycairo==1.28.0
pycparser==2.23
PyGObject==3.54.3
pykakasi==2.3.0
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
resemble-perth==1.0.1
s3tokenizer==0.2.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
soundfile==0.13.1
soxr==1.0.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.46.3
triton==3.2.0
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.34.0
wrapt==1.17.3
zipp==3.23.0
